studio/requirements.txt
langgraph
langchain
faiss-cpu
openai
requests
python-dotenv
langchain_openai
langchain
langchain-community
tiktoken

